+++ src/core.rs
@@ -1,2 +1,3 @@
 pub mod perlin;
 pub mod simplex;
+pub mod perlin_surflet;
